refactor: extract SolutionChangeWriter from CodeActionRunner
